correcion metodo cupon

+++ src/main/java/negocio/BOs/PedidoProgramadoBO.java
@@ -67,7 +67,7 @@ public void registrarPedido(PedidoProgramadoDTO dto)
     descuentoCalculado = subtotalCalculado *
             cuponValido.getDescuento().doubleValue();
 
-    idCuponAplicado = cuponValido.getIdCupon().intValue();
+    idCuponAplicado = cuponValido.getIdCupon();
     }
 
     if (Math.abs(subtotalCalculado - dto.getSubtotal()) > 0.01) {
